indexer/requirements.txt
web3==6.20.1
uvloop==0.21.0
python-dotenv==1.0.1
fastmcp==0.4.3